# contract/openapi.json
{
  "components": {
    "schemas": {
...
      "AppSettings": {
        "properties": {
          "allow_experimental_updates": {
            "default": false,
            "title": "Allow Experimental Updates",
            "type": "boolean"
          },
          "analytics_opt_in": {
            "default": true,
            "title": "Analytics Opt In",
            "type": "boolean"
          },
          "analytics_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analytics Token"
          },
          "anthropic_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anthropic Api Key"
          },
          "app_template_theme_override": {
            "anyOf": [
              {
                "enum": [
                  "light",
                  "dark"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Template Theme Override"
          },
          "auto_reveal_sub_agents": {
            "default": true,
            "title": "Auto Reveal Sub Agents",
            "type": "boolean"
          },
          "auto_select_mode_on_new_agent": {
            "default": false,
            "title": "Auto Select Mode On New Agent",
            "type": "boolean"
          },
          "browser_homepage": {
            "default": "https://www.google.com",
            "title": "Browser Homepage",
            "type": "string"
          },
          "claude_subscription_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claude Subscription Token"
          },
          "connection_mode": {
            "default": "own_key",
            "title": "Connection Mode",
            "type": "string"
          },
          "custom_providers": {
            "items": {
              "$ref": "#/components/schemas/CustomProvider"
            },
            "title": "Custom Providers",
            "type": "array"
          },
          "default_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Folder"
          },
          "default_max_turns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Max Turns"
          },
          "default_mode": {
            "default": "agent",
            "title": "Default Mode",
            "type": "string"
          },
          "default_model": {
            "default": "custom/maestro/maestro-fast",
            "title": "Default Model",
            "type": "string"
          },
          "default_system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "You are a personal AI assistant running inside Maestro.\n\n## Core Behavior\nAct, don't ask. When a tool can accomplish the task, call it immediately; do not describe what you would do, do not ask for confirmation, just execute. The user expects results, not plans.\nIf ANY available tool is relevant to the user's request, use it. Never respond with \"I can do X for you\" or \"Would you like me to...\"; just do it. A tool call is always better than a text explanation of what the tool would do.\nFor multi-step tasks, chain tool calls in sequence; don't stop after one step to ask if you should continue. Complete the entire task, then report the results.\nBe adaptable. If one approach fails, try a different tool or strategy instead of giving up or repeating the same action. Always stay focused on what the user actually wants to accomplish; their intent matters more than the specific method.\n\n## Tool Priority\n1. Connected MCP tools; fastest and most reliable. Use ToolSearch to discover what integrations are available if you're unsure.\n2. WebSearch / WebFetch; for general web lookups when no MCP tool fits.\n3. BrowserAgent; last resort, only for visual interaction with websites, filling forms, or tasks no other tool can handle.\n\n## Style\nDo not narrate routine tool calls; just call the tool.\nAfter tool calls complete, present the results directly. Do not recap which tools you called or why; the user can see tool calls in the UI.\nKeep responses brief and direct. Use plain language.\nIf you genuinely need clarification on something ambiguous, use the AskUserQuestion tool. Never ask questions inline in plain text.\n",
            "title": "Default System Prompt"
          },
          "default_thinking_level": {
            "default": "auto",
            "enum": [
              "off",
              "low",
              "medium",
              "high",
              "auto"
            ],
            "title": "Default Thinking Level",
            "type": "string"
          },
          "dev_mode": {
            "default": false,
            "title": "Dev Mode",
            "type": "boolean"
          },
          "dismissed_mcp_suggestions": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Dismissed Mcp Suggestions",
            "type": "object"
          },
          "expand_new_chats_in_dashboard": {
            "default": true,
            "title": "Expand New Chats In Dashboard",
            "type": "boolean"
          },
          "first_opened_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Opened At"
          },
          "gemini_subscription_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gemini Subscription Token"
          },
          "google_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Api Key"
          },
          "installation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installation Id"
          },
          "language": {
            "anyOf": [
              {
                "enum": [
                  "pt-BR",
                  "en"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "maestro_bearer_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maestro Bearer Token"
          },
          "maestro_proxy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maestro Proxy Url"
          },
          "new_agent_shortcut": {
            "default": "Meta+l",
            "title": "New Agent Shortcut",
            "type": "string"
          },
          "openai_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openai Api Key"
          },
          "openai_subscription_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openai Subscription Token"
          },
          "openrouter_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Api Key"
          },
          "preflight_enabled": {
            "default": true,
            "title": "Preflight Enabled",
            "type": "boolean"
          },
          "preflight_rollout_pct": {
            "default": 100,
            "title": "Preflight Rollout Pct",
            "type": "integer"
          },
          "provedor_ia_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provedor Ia Token"
          },
          "signin_method": {
            "anyOf": [
              {
                "enum": [
                  "google",
                  "stripe",
                  "email"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signin Method"
          },
          "theme": {
            "default": "light",
            "title": "Theme",
            "type": "string"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "user_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Email"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          },
          "user_referral_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Referral Source"
          },
          "user_use_case": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Use Case"
          },
          "zoom_sensitivity": {
            "default": 50,
            "title": "Zoom Sensitivity",
            "type": "number"
          }
        },
        "title": "AppSettings",
        "type": "object"
      },
...
      "CustomProvider": {
        "properties": {
          "api_key": {
            "default": "",
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "models": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Models",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "base_url"
        ],
        "title": "CustomProvider",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings": {
      "get": {
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "description": "Save only the fields the user changed, merged onto the CURRENT on-disk\nstate. The renderer sends a diff (not a stale full object), so a save can't\nclobber a field something else, an agent, an OAuth connect, changed\nunderneath it. Makes the lost-update unrepresentable: you can't overwrite a\nfield you never sent.",
        "operationId": "patch_settings_api_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Changes",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "update_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}